Atividade aula 4 de Java, sobre Scanner, janelas pelo JOption... dia 11/03/2026

+++ src/br/com/fiap/Exercício1Scanner.java
@@ -10,10 +10,10 @@ public class Exercício1Scanner {
         try {
             scan = new Scanner(System.in);
             System.out.println("Dígite a nota das 4 provas.");
-            p1 = scan.nextInt();
-            p2 = scan.nextInt();
-            p3 = scan.nextInt();
-            p4 = scan.nextInt();
+            p1 = scan.nextFloat();
+            p2 = scan.nextFloat();
+            p3 = scan.nextFloat();
+            p4 = scan.nextFloat();
             media = p1 + p2 + p3 + p4 /4;
             System.out.println("Prova 1: " + p1 + "\nProva 2: " + p2 + "\nProva 3: " + p3 + "\nProva 4: " + p4 + "\nMédia: " + media);
         } catch (Exception e) {
